Shadow Mapping with LWJGL and shaders
Shadow mapping with LWJGL and shaders

After looking around for days I couldn't find a working example of
shadow mapping written in Java with LWJGL and shaders. Now, when I
finally managed to figure it out from tutorials in other languages I'd
like to share it to help other learners.

Disclaimer: I'm an openGL beginner and the code probably isn't ther best
written but it works and could be used as a guideline.

Notice that the light is a point light and the shadows are made with
orthogonal projection (used for directional light).